Separate a few more parts
