+++ EpikCord/__init__.py
@@ -7,13 +7,12 @@ import asyncio
 import datetime
 import re
 import socket
+from .flags import *
 import struct
-from abc import abstractmethod
 from base64 import b64encode
 from collections import defaultdict
 from logging import getLogger
 from sys import platform
-from time import perf_counter_ns
 from typing import (
     Optional,
     List,
@@ -22,8 +21,7 @@ from typing import (
     TypeVar,
     Callable,
     Tuple,
-    Type,
-    TYPE_CHECKING,
+    Type
 )
 
 from aiohttp import ClientSession, ClientResponse, ClientWebSocketResponse
@@ -122,7 +120,6 @@ class Connection:
         self.show_activity: bool = data["show_activity"]
         self.visibility: VisibilityType = VisibilityType(data["visibility"])
 
-
 class AuthorizationInformation:
     def __init__(self, data: dict):
         self.application: Application = Application(data["application"])
@@ -198,74 +195,6 @@ class UserClient:
 
 Localisation = Localization
 
-
-class CommandHandler:
-    def __init__(self):
-        self.commands: Dict[
-            str, Union[ClientSlashCommand, ClientUserCommand, ClientMessageCommand]
-        ] = {}
-
-    def command(
-        self,
-        *,
-        name: Optional[str] = None,
-        description: str = None,
-        guild_ids: Optional[List[str]] = None,
-        options: Optional[List[AnyOption]] = None,
-        name_localizations: Optional[List[Localization]] = None,
-        description_localizations: Optional[List[Localization]] = None,
-        name_localisations: Optional[List[Localization]] = None,
-        description_localisations: Optional[List[Localization]] = None,
-    ):
-        name_localization = self.utils.match_mixed(
-            name_localizations, name_localisations
-        )
-        description_localization = self.utils.match_mixed(
-            description_localizations, description_localisations
-        )
-
-        def register_slash_command(func):
-            desc = description or func.__doc__
-            if not desc:
-                raise TypeError(
-                    f"Command with {name or func.__name__} has no description. This is required."
-                )
-
-            command = ClientSlashCommand(
-                name=name or func.__name__,
-                description=desc,
-                guild_ids=guild_ids or [],
-                options=options or [],
-                callback=func,
-                name_localization=name_localization,
-                description_localization=description_localization,
-            )
-
-            self.commands[command.name] = command
-            return command
-
-        return register_slash_command
-
-    def user_command(self, name: Optional[str] = None):
-        def register_slash_command(func):
-
-            results = ClientUserCommand(callback=func, name=name or func.__name__)
-
-            self.commands[name] = results
-            return results
-
-        return register_slash_command
-
-    def message_command(self, name: Optional[str] = None):
-        def register_slash_command(func):
-
-            results = ClientMessageCommand(callback=func, name=name or func.__name__)
-
-            self.commands[name] = results
-            return results
-
-        return register_slash_command
-
 class UnavailableGuild:
     """
     The class representation of an UnavailableGuild.
@@ -306,341 +235,6 @@ class VoiceRegion:
         self.optimal: bool = data["optimal"]
         self.deprecated: bool = data["deprecated"]
         self.custom: bool = data["custom"]
-
-class WebsocketClient(EventHandler):
-    def __init__(self, token: str, intents: int):
-        super().__init__()
-
-        self.token = token
-        if not token:
-            raise TypeError("Missing token.")
-
-        if isinstance(intents, int):
-            self.intents = Intents(intents)
-        elif isinstance(intents, Intents):
-            self.intents = intents
-
-        self._closed = True
-        self.heartbeats = []
-
-        self.interval = None  # How frequently to heartbeat
-        self.session_id = None
-        self.sequence = None
-
-    async def change_presence(self, *, presence: Optional[Presence]):
-        payload = {"op": GatewayOpcode.PRESENCE_UPDATE, "d": presence.to_dict()}
-        await self.send_json(payload)
-
-    async def heartbeat(self, forced: Optional[bool] = None):
-        if forced:
-            return await self.send_json(
-                {"op": GatewayOpcode.HEARTBEAT, "d": self.sequence or "null"}
-            )
-
-        if self.interval:
-            await self.send_json(
-                {"op": GatewayOpcode.HEARTBEAT, "d": self.sequence or "null"}
-            )
-            self.heartbeat_time = perf_counter_ns()
-            await asyncio.sleep(self.interval / 1000)
-            logger.debug("Sent a heartbeat!")
-
-    async def request_guild_members(
-        self,
-        guild_id: int,
-        *,
-        query: Optional[str] = None,
-        limit: Optional[int] = None,
-        presences: Optional[bool] = None,
-        user_ids: Optional[List[str]] = None,
-        nonce: Optional[str] = None,
-    ):
-        payload = {
-            "op": GatewayOpcode.REQUEST_GUILD_MEMBERS,
-            "d": {"guild_id": guild_id},
-        }
-
-        if query:
-            payload["d"]["query"] = query
-
-        if limit:
-            payload["d"]["limit"] = limit
-
-        if presences:
-            payload["d"]["presences"] = presences
-
-        if user_ids:
-            payload["d"]["user_ids"] = user_ids
-
-        if nonce:
-            payload["d"]["nonce"] = nonce
-
-        await self.send_json(payload)
-
-    async def reconnect(self):
-        await self.close()
-        await self.connect()
-        await self.identify()
-        await self.resume()
-
-    async def handle_close(self):
-        if self.ws.close_code == GatewayCECode.DisallowedIntents:
-            raise DisallowedIntents(
-                "You cannot use privileged intents with this token, go to "
-                "the developer portal and allow the privileged intents "
-                "needed. "
-            )
-        elif self.ws.close_code == 1006:
-            await self.resume()
-        elif self.ws.close_code == GatewayCECode.AuthenticationFailed:
-            raise InvalidToken("The token you provided is invalid.")
-        elif self.ws.close_code == GatewayCECode.RateLimited:
-            raise Ratelimited429(
-                "You've been rate limited. Try again in a few minutes."
-            )
-        elif self.ws.close_code == GatewayCECode.ShardingRequired:
-            raise ShardingRequired("You need to shard the bot.")
-        elif self.ws.close_code == GatewayCECode.InvalidAPIVersion:
-            raise DeprecationWarning(
-                "The gateway you're connecting to is deprecated and does not "
-                "work, upgrade EpikCord.py. "
-            )
-        elif self.ws.close_code == GatewayCECode.InvalidIntents:
-            raise InvalidIntents("The intents you provided are invalid.")
-        elif self.ws.close_code == GatewayCECode.UnknownError:
-            await self.resume()
-        elif self.ws.close_code == GatewayCECode.UnknownOpcode:
-            logger.critical(
-                "EpikCord.py sent an invalid OPCODE to the Gateway. "
-                "Report this immediately. "
-            )
-            await self.resume()
-        elif self.ws.close_code == GatewayCECode.DecodeError:
-            logger.critical(
-                "EpikCord.py sent an invalid payload to the Gateway."
-                " Report this immediately. "
-            )
-            await self.resume()
-        elif self.ws.close_code == GatewayCECode.NotAuthenticated:
-            logger.critical(
-                "EpikCord.py has sent a payload prior to identifying."
-                " Report this immediately. "
-            )
-
-        elif self.ws.close_code == GatewayCECode.AlreadyAuthenticated:
-            logger.critical(
-                "EpikCord.py tried to authenticate again." " Report this immediately. "
-            )
-            await self.resume()
-        elif self.ws.close_code == GatewayCECode.InvalidSequence:
-            logger.critical(
-                "EpikCord.py sent an invalid sequence number."
-                " Report this immediately."
-            )
-            await self.resume()
-        elif self.ws.close_code == GatewayCECode.SessionTimeout:
-            logger.critical("Session timed out.")
-            await self.resume()
-        else:
-            raise ClosedWebSocketConnection(
-                f"Connection has been closed with code {self.ws.close_code}"
-            )
-
-    async def send_json(self, json: dict):
-        await self.ws.send_json(json)
-        logger.debug(f"Sent {json} to the Websocket Connection to Discord.")
-
-    async def connect(self):
-        res = await self.http.get("/gateway")
-        data = await res.json()
-        url = data["url"]
-        self.ws = await self.http.ws_connect(
-            f"{url}?v=10&encoding=json&compress=zlib-stream"
-        )
-        self._closed = False
-        await self.handle_events()
-
-    async def resume(self):
-        logger.critical("Reconnecting...")
-        await self.connect()
-        await self.send_json(
-            {
-                "op": GatewayOpcode.RESUME,
-                "d": {
-                    "seq": self.sequence,
-                    "session_id": self.session_id,
-                    "token": self.token,
-                },
-            }
-        )
-        self._closed = False
-
-    async def identify(self):
-        payload = {
-            "op": GatewayOpcode.IDENTIFY,
-            "d": {
-                "token": self.token,
-                "intents": self.intents.value,
-                "properties": {
-                    "os": platform,
-                    "browser": "EpikCord.py",
-                    "device": "EpikCord.py",
-                },
-            },
-        }
-        if self.presence:
-            payload["d"]["presence"] = self.presence.to_dict()
-        return await self.send_json(payload)
-
-    async def close(self) -> None:
-        if self._closed:
-            return
-
-        # for voice in self.voice_clients:
-        #     try:
-        #         await voice.disconnect(force=True)
-        #     except Exception:
-        #         # if an error happens during disconnects, disregard it.
-        #         pass
-
-        if self.ws is not None and not self.ws.closed:
-            await self.ws.close(code=4000)
-
-        if self.http is not None and not self.http.closed:
-            await self.http.close()
-
-        self._closed = True
-
-    def login(self):
-
-        loop = asyncio.get_event_loop()
-
-        async def runner():
-            try:
-                await self.connect()
-            finally:
-                if not self._closed:
-                    await self.close()
-
-        def stop_loop_on_completion(f: asyncio.Future):
-            loop.stop()
-
-        future = asyncio.ensure_future(runner(), loop=loop)
-        future.add_done_callback(stop_loop_on_completion)
-
-        try:
-            loop.run_forever()
-        except KeyboardInterrupt:
-            pass
-        finally:
-            future.remove_done_callback(stop_loop_on_completion)
-            self.utils.cleanup_loop(loop)
-
-
-class BaseCommand:
-    def __init__(self):
-        self.checks: List[Check] = []
-
-    def is_slash_command(self):
-        return self.type == 1
-
-    def is_user_command(self):
-        return self.type == 2
-
-    def is_message_command(self):
-        return self.type == 3
-
-    @abstractmethod
-    @property
-    def type(self):
-        ...
-
-
-class ClientUserCommand(BaseCommand):
-    """
-    A class to represent a User Command that the Client owns.
-
-    Attributes:
-    -----------
-        * name The name set for the User Command
-        * callback: callable The function to call for the User Command
-        (Passed in by the library)
-
-    Parameters:
-    -----------
-    All parameters follow the documentation of the Attributes accordingly
-        * name
-        * callback
-    """
-
-    def __init__(self, *, name: str, callback: Callable):
-        super().__init__()
-        self.name: str = name
-        self.callback: Callable = callback
-
-    @property
-    def type(self):
-        return 2
-
-
-class ClientSlashCommand(BaseCommand):
-    def __init__(
-        self,
-        *,
-        name: str,
-        description: str,
-        callback: Callable,
-        guild_ids: Optional[List[str]] = None,
-        options: Optional[List[AnyOption]] = None,
-        name_localization: Optional[Localization] = None,
-        description_localization: Optional[str] = None,
-    ):
-        super().__init__()
-        self.name: str = name
-        self.description: str = description
-        self.name_localizations: Optional[Localization] = name_localization
-        self.description_localizations: Optional[
-            Localization
-        ] = description_localization
-        self.callback: Callable = callback
-        self.guild_ids: Optional[List[str]] = guild_ids or []
-        self.options: Optional[List[AnyOption]] = options or []
-        self.autocomplete_options: dict = {}
-
-    @property
-    def type(self):
-        return 1
-
-    def option_autocomplete(self, option_name: str):
-        def wrapper(func):
-            self.autocomplete_options[option_name] = func
-
-        return wrapper
-
-    def to_dict(self):
-        payload = {
-            "name": self.name,
-            "description": self.description,
-            "type": self.type,
-            "options": [option.to_dict() for option in self.options],
-        }
-
-        if self.name_localizations:
-            payload["name_localizations"] = [
-                l.to_dict() for l in self.name_localizations
-            ]
-        if self.description_localizations:
-            payload["description_localizations"] = [
-                l.to_dict() for l in self.description_localizations
-            ]
-        return payload
-
-
-class ClientMessageCommand(ClientUserCommand):
-    @property
-    def type(self):
-        return 3
-
 
 class Overwrite:
     def __init__(self, data: dict):
@@ -2105,29 +1699,6 @@ class Integration:
             Application(data.get("application")) if data.get("application") else None
         )
 
-
-# TODO Adapt this to subclass Flags
-class SystemChannelFlags:
-    def __init__(self, *, value: Optional[int] = None):
-        self.value: int = value
-
-    @property
-    def suppress_join_notifications(self):
-        self.value += 1 << 0
-
-    @property
-    def suppress_premium_subscriptions(self):
-        self.value += 1 << 1
-
-    @property
-    def suppress_guild_reminder_notifications(self):
-        self.value += 1 << 2
-
-    @property
-    def suppress_join_notification_replies(self):
-        self.value += 1 << 3
-
-
 class Guild:
     def __init__(self, client: Client, data: dict):
         self.client = client
@@ -2822,116 +2393,6 @@ class Webhook:  # Not used for making webhooks.
         self.url: Optional[str] = data.get("url")
 
 
-class Flag:
-    if TYPE_CHECKING:
-        class_flags: "dict[str, Any]"
-
-    def __init_subclass__(cls) -> None:
-        cls.class_flags = {k: v for k, v in cls.__dict__.items() if isinstance(v, int)}
-        return cls
-
-    def __init__(self, value: int = 0, **kwargs):
-        self.value = value
-        self.turned_on: "list[str]" = [k for k, a in kwargs.items() if a]
-
-        for k, v in self.class_flags.items():
-            if v & value and k not in self.turned_on:
-                self.turned_on.append(k)
-
-        self.calculate_from_turned()
-
-    def calculate_from_turned(self):
-        value = 0
-        for key, flag in self.class_flags.items():
-            if key in self.class_flags:
-                value |= flag
-        self.value = value
-
-    def __getattribute__(self, __name: str) -> Any:
-        original = super().__getattribute__
-        if __name in original("class_flags"):
-            return __name in original("turned_on")
-        return original(__name)
-
-    def __setattr__(self, __name: str, __value: Any) -> None:
-        if __name not in self.class_flags:
-            return super().__setattr__(__name, __value)
-        if __value and __name not in self.turned_on:
-            self.turned_on.append(__name)
-        elif not __value and __name in self.turned_on:
-            self.turned_on.remove(__name)
-        self.calculate_from_turned()
-
-    @classmethod
-    def all(cls):
-        return cls(**{k: True for k in cls.class_flags})
-
-
-class Intents(Flag):
-    guilds = 1 << 0
-    members = 1 << 1
-    bans = 1 << 2
-    emojis_and_stickers = 1 << 3
-    integrations = 1 << 4
-    webhooks = 1 << 5
-    invites = 1 << 6
-    voice_states = 1 << 7
-    presences = 1 << 8
-
-    guild_messages = 1 << 9
-    guild_message_reactions = 1 << 10
-    guild_message_typing = 1 << 11
-
-    direct_messages = 1 << 12
-    direct_message_reactions = 1 << 13
-    direct_message_typing = 1 << 14
-
-    message_content = 1 << 15
-    scheduled_event = 1 << 16
-
-
-class Permissions(Flag):
-    create_instant_invite = 1 << 0
-    kick_members = 1 << 1
-    ban_members = 1 << 2
-    administrator = 1 << 3
-    manage_channels = 1 << 4
-    manage_guild = 1 << 5
-    add_reactions = 1 << 6
-    view_audit_log = 1 << 7
-    priority_speaker = 1 << 8
-    stream = 1 << 9
-    read_messages = 1 << 10
-    send_messages = 1 << 11
-    send_tts_messages = 1 << 12
-    manage_messages = 1 << 13
-    embed_links = 1 << 14
-    attach_files = 1 << 15
-    read_message_history = 1 << 16
-    mention_everyone = 1 << 17
-    use_external_emojis = 1 << 18
-    connect = 1 << 20
-    speak = 1 << 21
-    mute_members = 1 << 22
-    deafen_members = 1 << 23
-    move_members = 1 << 24
-    use_voice_activation = 1 << 25
-    change_nickname = 1 << 26
-    manage_nicknames = 1 << 27
-    manage_roles = 1 << 28
-    manage_webhooks = 1 << 29
-    manage_emojis_and_stickers = 1 << 30
-    use_application_commands = 1 << 31
-    request_to_speak = 1 << 32
-    manage_events = 1 << 33
-    manage_threads = 1 << 34
-    create_public_threads = 1 << 35
-    create_private_threads = 1 << 36
-    use_external_stickers = 1 << 37
-    send_messages_in_threads = 1 << 38
-    start_embedded_activities = 1 << 39
-    moderator_members = 1 << 40
-
 class Paginator:
     def __init__(self, *, pages: List[Embed]):
         self.current_index: int = 0
@@ -3501,36 +2962,6 @@ class ShardManager(CommandHandler, EventHandler):
 
         loop = asyncio.get_event_loop()
         loop.run_until_complete(wrapper())
-
-
-class Check:
-    def __init__(self, callback):
-        self.callback = callback
-        self.success_callback = self.default_success
-        self.failure_callback = self.default_failure
-
-    def success(self, callback: Optional[Callable] = None):
-        self.success_callback = callback or self.default_success
-
-    def failure(self, callback: Optional[Callable] = None):
-        self.failure_callback = callback or self.default_failure
-
-    async def default_success(self, interaction):
-        logger.info(
-            f"{interaction.author.username} ({interaction.author.id}) passed "
-            f"the check {self.command_callback.__name__}. "
-        )
-
-    async def default_failure(self, interaction):
-        logger.critical(
-            f"{interaction.author.username} ({interaction.author.id}) failed "
-            f"the check {self.command_callback.__name__}. "
-        )
-        raise FailedCheck(
-            f"{interaction.author.username} ({interaction.author.id}) failed "
-            f"the check {self.command_callback.__name__}. "
-        )
-
 
 class CommandUtils:
     @staticmethod
