+++ EpikCord/commands/__init__.py
@@ -1,0 +1,3 @@
+from .checks import *
+from .command_handler import *
+from .commands import *
