+++ EpikCord/flags/__init__.py
@@ -1,0 +1,2 @@
+from .intents import *
+from .system_channel_flags import *
