+++ EpikCord/flags/intents.py
@@ -1,0 +1,23 @@
+from .flag import Flag
+
+class Intents(Flag):
+    guilds = 1 << 0
+    members = 1 << 1
+    bans = 1 << 2
+    emojis_and_stickers = 1 << 3
+    integrations = 1 << 4
+    webhooks = 1 << 5
+    invites = 1 << 6
+    voice_states = 1 << 7
+    presences = 1 << 8
+
+    guild_messages = 1 << 9
+    guild_message_reactions = 1 << 10
+    guild_message_typing = 1 << 11
+
+    direct_messages = 1 << 12
+    direct_message_reactions = 1 << 13
+    direct_message_typing = 1 << 14
+
+    message_content = 1 << 15
+    scheduled_event = 1 << 16
