+++ EpikCord/flags/system_channel_flags.py
@@ -1,0 +1,7 @@
+from .flag import Flag
+
+class SystemChannelFlags(Flag):
+    suppress_join_notifications = 1 << 0
+    suppress_premium_subscriptions = 1 << 1
+    suppress_guild_reminder_notifications = 1 << 2
+    suppress_join_notification_replies = 1 << 3
